Simplifying syntax for lift and drag coefficients Adding a base class to make lift definitions easier for F9 S1 varients Moving Pi constants to a central area Moving Yaw and Roll to the spacecraft level
