+++ src/SpaceSim/Physics/Constants.cs
@@ -1,0 +1,14 @@
+﻿namespace SpaceSim.Physics
+{
+    public static class Constants
+    {
+        public const double GravitationConstant = 6.67384e-11;
+
+        public const double SpeedOfLight = 2.998e8;
+
+        public const double SpeedLightSquared = 8.988004e16;
+
+        public const double TwoPi = 6.283185307179586476925286766559;
+        public const double PiOverTwo = 1.5707963267948966192313216916398;
+    }
+}
