--- src/SpaceSim/Physics/FlightGlobals.cs
@@ -1,11 +1,0 @@
-﻿namespace SpaceSim.Physics
-{
-    public static class FlightGlobals
-    {
-        public const double GRAVITATION_CONSTANT = 6.67384e-11;
-
-        public const double SPEED_OF_LIGHT = 2.998e8;
-
-        public const double SPEED_LIGHT_SQUARED = 8.988004e16;
-    }
-}
