+++ src/SpaceSim/Physics/IGravitationalBody.cs
@@ -13,10 +13,11 @@ namespace SpaceSim.Physics
 
         void ResolveGravitation(IPhysicsBody other);
 
+        double GetRelativePitch();
         double GetRelativeAltitude();
+
         DVector2 GetRelativeVelocity();
         DVector2 GetRelativeAcceleration();
-        double GetRelativePitch();
 
         void SetGravitationalParent(IMassiveBody parent);
     }
