+++ src/SpaceSim/Spacecrafts/Falcon9SSTO/F9SSTO.cs
@@ -2,64 +2,23 @@
 using System.Drawing;
 using SpaceSim.Engines;
 using SpaceSim.Physics;
+using SpaceSim.Spacecrafts.FalconCommon;
 using VectorMath;
 
 namespace SpaceSim.Spacecrafts.Falcon9SSTO
 {
-    sealed class F9SSTO : SpaceCraftBase
+    sealed class F9SSTO : F9S1Base
     {
         public override string CraftName { get { return "F9 S1"; } }
+        public override string CommandFileName { get { return "F9SSTO.xml"; } }
 
         public override double DryMass { get { return 21600; } }
 
         public override double Width { get { return 4.11; } }
         public override double Height { get { return 47.812188; } }
 
-        public override AeroDynamicProperties GetAeroDynamicProperties { get { return AeroDynamicProperties.ExtendsFineness; } }
-
-        public override double FormDragCoefficient
-        {
-            get
-            {
-                double baseCd = GetBaseCd(0.15);
-                double alpha = GetAlpha();
-                double cosAlpha = Math.Cos(alpha);
-                double Cd = Math.Abs(baseCd * cosAlpha);
-
-                return Cd;
-            }
-        }
-
-        public override double LiftCoefficient
-        {
-            get
-            {
-                double baseCd = GetBaseCd(0.6);
-                double alpha = GetAlpha();
-                double sinAlpha = Math.Sin(alpha * 2);
-                return baseCd * sinAlpha;
-            }
-        }
-
-        public override double CrossSectionalArea { get { return Math.PI * Math.Pow(Width / 2, 2); } }
-
-        public override double ExposedSurfaceArea
-        {
-            get
-            {
-                // A = 2πrh + πr2
-                return 2 * Math.PI * (Width / 2) * Height + CrossSectionalArea;
-            }
-        }
-
-        public override double LiftingSurfaceArea { get { return Width * Height; } }
-
-        public override Color IconColor { get { return Color.White; } }
-
-        public override string CommandFileName { get { return "F9SSTO.xml"; } }
-
         public F9SSTO(string craftDirectory, DVector2 position, DVector2 velocity)
-            : base(craftDirectory, position, velocity, 409500, "Textures/f9ssto.png")
+            : base(craftDirectory, position, velocity, 409500, "Textures/f9ssto.png", null)
         {
             StageOffset = new DVector2(0, 25.5);
 
@@ -74,9 +33,5 @@ namespace SpaceSim.Spacecrafts.Falcon9SSTO
                 Engines[i] = new Merlin1D(i, this, offset);
             }
         }
-
-        public void DeployFins() { }
-
-        public void DeployLegs() { }
     }
 }
