+++ src/SpaceSim/Spacecrafts/FalconHeavy/FHBooster.cs
@@ -2,13 +2,19 @@
 using System.Drawing;
 using SpaceSim.Engines;
 using SpaceSim.Physics;
+using SpaceSim.Spacecrafts.FalconCommon;
 using VectorMath;
 
 namespace SpaceSim.Spacecrafts.FalconHeavy
 {
-    sealed class FHBooster : SpaceCraftBase
+    sealed class FHBooster : F9S1Base
     {
         public override string CraftName { get { return "FH Booster " + Id; } }
+
+        public override string CommandFileName
+        {
+            get { return Id == 1 ? "FHLeftBooster.xml" : "FHRightBooster.xml"; }
+        }
 
         public int Id { get; private set; }
 
@@ -17,64 +23,12 @@ namespace SpaceSim.Spacecrafts.FalconHeavy
         public override double Width { get { return 4.11; } }
         public override double Height { get { return 44.6; } }
 
-        public override AeroDynamicProperties GetAeroDynamicProperties { get { return AeroDynamicProperties.ExtendsFineness; } }
-
-        public override double FormDragCoefficient
-        {
-            get
-            {
-                double baseCd = GetBaseCd(0.4);
-                double alpha = GetAlpha();
-                double cosAlpha = Math.Cos(alpha);
-                double Cd = Math.Abs(baseCd * cosAlpha);
-
-                return Cd;
-            }
-        }
-
-        public override double LiftCoefficient
-        {
-            get
-            {
-                double baseCd = GetBaseCd(0.6);
-                double alpha = GetAlpha();
-                double sinAlpha = Math.Sin(alpha * 2);
-                return baseCd * sinAlpha;
-            }
-        }
-
-        public override double CrossSectionalArea { get { return Math.PI * Math.Pow(Width / 2, 2); } }
-        public override double ExposedSurfaceArea
-        {
-            get
-            {
-                // A = 2πrh + πr2
-                return 2 * Math.PI * (Width / 2) * Height + CrossSectionalArea;
-            }
-        }
-
-        public override double LiftingSurfaceArea { get { return Width * Height; } }
-
-        public override Color IconColor { get { return Color.White; } }
-
-        public override string CommandFileName
-        {
-            get { return Id == 1 ? "FHLeftBooster.xml" : "FHRightBooster.xml"; }
-        }
-
         public FHBooster(string craftDirectory, int id, DVector2 position, DVector2 velocity)
-            : base(craftDirectory, position, velocity, 398887, "Textures/fhBooster" + id + ".png")
+            : base(craftDirectory, position, velocity, 398887, "Textures/fhBooster" + id + ".png", null)
         {
             Id = id;
 
-            if (Id == 1)
-            {
-                StageOffset = new DVector2(-4, 1.5);   
-            }
-            else
-            {
-                StageOffset = new DVector2(4, 1.5);
-            }
+            StageOffset = Id == 1 ? new DVector2(-4, 1.5) : new DVector2(4, 1.5);
 
             Engines = new IEngine[9];
